fixes and tests

+++ src/instrument.h
@@ -54,6 +54,6 @@ class Instrument : public Engine {
 u16 updateCurrentFrame(long& current_frame_old, long& current_frame_new,
                        u16 offset_delta_old, const u32* relocation_table);
 u8 parametersSlots(const char* method_sig);
-u8 instructionBytes(const u8* code, u32 index);
+u32 instructionBytes(const u8* code, u32 index);
 
 #endif // _INSTRUMENT_H
